# app/try_on_api.py
"""REST routes for per-model virtual try-on configuration and photo compose."""

from __future__ import annotations

import os
from pathlib import Path
from typing import Annotated

from fastapi import APIRouter, Depends, File, Form, HTTPException, UploadFile
from fastapi.responses import JSONResponse

from .auth_deps import resolve_user_id

from .job_manager import JobManager
from .photo_compose_models import PhotoComposeResponse
from .photo_compose_service import PhotoComposeService
from .runtime_settings import RuntimeSettings, SettingsStore
from .try_on_config import (
    HangerPointBody,
    ModelPoint3D,
    ModelTryOnConfig,
    ModelTryOnConfigPut,
    TryOnConfigStore,
    UserTryOnCalibration,
    UserTryOnCalibrationPut,
)

router = APIRouter(tags=["try-on"])

_store: TryOnConfigStore | None = None
_compose_service: PhotoComposeService | None = None
_job_manager: JobManager | None = None
_settings_store: SettingsStore | None = None
_root_dir: Path | None = None
_uploads_dir: Path | None = None


def init_try_on_api(
    *,
    job_manager: JobManager,
    settings_store: SettingsStore,
    root_dir: Path,
    uploads_dir: Path,
) -> None:
    global _store, _compose_service, _job_manager, _settings_store, _root_dir, _uploads_dir
    _job_manager = job_manager
    _settings_store = settings_store
    _root_dir = root_dir
    _uploads_dir = uploads_dir
    _store = TryOnConfigStore(job_manager.db_path)
    _compose_service = PhotoComposeService(
        db_path=job_manager.db_path,
        root_dir=root_dir,
        job_manager=job_manager,
        settings_store=settings_store,
        uploads_dir=uploads_dir,
    )


def get_try_on_store() -> TryOnConfigStore:
    if _store is None:
        raise RuntimeError("try_on_api not initialized")
    return _store


def get_photo_compose_service() -> PhotoComposeService:
    if _compose_service is None:
        raise RuntimeError("try_on_api not initialized")
    return _compose_service


_FACE_MIME = frozenset({"image/jpeg", "image/jpg", "image/png"})


def _output_dir() -> Path:
    assert _settings_store is not None and _root_dir is not None
    settings = _settings_store.load()
    return _root_dir / settings.output_dir_name


def _model_output_exists(job_id: str) -> bool:
    out = _output_dir()
    if not out.is_dir():
        return False
    return (out / f"{job_id}.glb").is_file() or (out / f"{job_id}.ply").is_file()


def _assert_model_job_exists(job_id: str) -> None:
    assert _job_manager is not None
    job = _job_manager.get_job(job_id)
    if job is None and not _model_output_exists(job_id):
        raise HTTPException(status_code=404, detail="Job not found")


def _require_authenticated_user_id(user_id: str | None = Depends(resolve_user_id)) -> str:
    if not user_id:
        raise HTTPException(
            status_code=401,
            detail="Authentication required. Send Authorization: Bearer <token> or X-User-Id.",
        )
    return user_id


def _enforce_owner_access(job_id: str, user_id: str | None) -> None:
    """When a row has owner_user_id, require matching authenticated user (403)."""
    store = get_try_on_store()
    owner = store.get_owner(job_id)
    if owner and owner != user_id:
        raise HTTPException(status_code=403, detail="Not authorized for this model")


def _resolve_owner_on_write(job_id: str, user_id: str | None) -> str | None:
    store = get_try_on_store()
    existing_owner = store.get_owner(job_id)
    if existing_owner:
        return existing_owner
    if user_id and os.getenv("APP_TRY_ON_BIND_OWNER", "1").strip().lower() not in ("0", "false", "no"):
        return user_id
    return None


@router.get("/models/{job_id}/try-on", response_model=ModelTryOnConfig)
def get_model_try_on(job_id: str, user_id: str | None = Depends(resolve_user_id)) -> ModelTryOnConfig:
    _assert_model_job_exists(job_id)
    _enforce_owner_access(job_id, user_id)
    cfg = get_try_on_store().get(job_id)
    if cfg is None:
        raise HTTPException(status_code=404, detail="Try-on config not found for this model")
    return cfg


@router.put("/models/{job_id}/try-on", response_model=ModelTryOnConfig)
def put_model_try_on(
    job_id: str,
    body: ModelTryOnConfigPut,
    user_id: str | None = Depends(resolve_user_id),
) -> ModelTryOnConfig:
    _assert_model_job_exists(job_id)
    _enforce_owner_access(job_id, user_id)
    store = get_try_on_store()
    existing = store.get(job_id)
    if existing is None and body.hanger_point is None:
        raise HTTPException(
            status_code=400,
            detail="hanger_point is required when creating try-on config.",
        )
    try:
        owner = _resolve_owner_on_write(job_id, user_id)
        return store.upsert(job_id, body, require_hanger=existing is None, owner_user_id=owner)
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from None


@router.get("/models/{job_id}/hanger-point", response_model=ModelPoint3D)
def get_hanger_point(job_id: str, user_id: str | None = Depends(resolve_user_id)) -> ModelPoint3D:
    _assert_model_job_exists(job_id)
    _enforce_owner_access(job_id, user_id)
    cfg = get_try_on_store().get(job_id)
    if cfg is None:
        raise HTTPException(status_code=404, detail="Hanger point not found for this model")
    return cfg.hanger_point


@router.put("/models/{job_id}/hanger-point", response_model=ModelPoint3D)
def put_hanger_point(
    job_id: str,
    body: HangerPointBody,
    user_id: str | None = Depends(resolve_user_id),
) -> ModelPoint3D:
    _assert_model_job_exists(job_id)
    _enforce_owner_access(job_id, user_id)
    store = get_try_on_store()
    try:
        owner = _resolve_owner_on_write(job_id, user_id)
        cfg = store.upsert_hanger_only(
            job_id,
            ModelPoint3D(x=body.x, y=body.y, z=body.z),
            owner_user_id=owner,
        )
        return cfg.hanger_point
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from None


@router.get("/users/me/try-on-calibration", response_model=UserTryOnCalibration)
def get_user_try_on_calibration(
    user_id: str = Depends(_require_authenticated_user_id),
) -> UserTryOnCalibration:
    cfg = get_try_on_store().get_user_calibration(user_id)
    if cfg is None:
        raise HTTPException(status_code=404, detail="Try-on calibration not found for this user")
    return cfg


@router.put("/users/me/try-on-calibration", response_model=UserTryOnCalibration)
def put_user_try_on_calibration(
    body: UserTryOnCalibrationPut,
    user_id: str = Depends(_require_authenticated_user_id),
) -> UserTryOnCalibration:
    return get_try_on_store().upsert_user_calibration(user_id, body)


@router.post(
    "/try-on/photo-compose",
    response_model=PhotoComposeResponse,
    status_code=202,
    responses={
        200: {"description": "Synchronous completion (rare)"},
        202: {"description": "Compose job accepted; poll GET by compose_id"},
        503: {"description": "AI provider not configured"},
    },
)
async def post_photo_compose(
    face_image: UploadFile = File(..., description="Side/profile face photo (JPEG or PNG)"),
    job_id: str = Form(..., description="Gallery earring model id"),
    placement_x: float = Form(..., description="Normalized X in native image (0=left, 1=right)"),
    placement_y: float = Form(..., description="Normalized Y in native image (0=top, 1=bottom)"),
    image_width: int = Form(..., gt=0, description="Native pixel width of face_image"),
    image_height: int = Form(..., gt=0, description="Native pixel height of face_image"),
    placement_side: str = Form(default="auto", description="left | right | auto"),
    prompt: str | None = Form(default=None, description="Optional extra instruction (max 2000 chars)"),
    user_id: str | None = Depends(resolve_user_id),
) -> JSONResponse:
    if not (0.0 <= placement_x <= 1.0):
        raise HTTPException(status_code=400, detail="placement_x must be between 0 and 1")
    if not (0.0 <= placement_y <= 1.0):
        raise HTTPException(status_code=400, detail="placement_y must be between 0 and 1")
    side = (placement_side or "auto").strip().lower()
    if side not in ("left", "right", "auto"):
        raise HTTPException(status_code=400, detail="placement_side must be left, right, or auto")
    if prompt is not None and len(prompt) > 2000:
        raise HTTPException(status_code=400, detail="prompt must be at most 2000 characters")

    mime = (face_image.content_type or "").split(";")[0].strip().lower()
    if mime and mime not in _FACE_MIME:
        raise HTTPException(status_code=400, detail="face_image must be JPEG or PNG")

    svc = get_photo_compose_service()
    max_bytes = svc.max_upload_bytes()
    raw = await face_image.read()
    if len(raw) > max_bytes:
        raise HTTPException(status_code=413, detail=f"face_image exceeds {max_bytes // (1024 * 1024)} MB limit")
    if not raw:
        raise HTTPException(status_code=400, detail="face_image is empty")

    gemini_ok = bool(os.getenv("GEMINI_API_KEY", "").strip())
    dev_mock = os.getenv("PHOTO_COMPOSE_DEV_MOCK", "").strip().lower() in ("1", "true", "yes")
    if not gemini_ok and not dev_mock:
        raise HTTPException(
            status_code=503,
            detail="Photo compose AI is not configured (set GEMINI_API_KEY on the API server)",
        )

    try:
        svc.assert_model_exists(job_id)
    except KeyError:
        raise HTTPException(status_code=404, detail="Job not found") from None

    try:
        svc.resolve_model_image_url(job_id)
    except ValueError as exc:
        raise HTTPException(status_code=422, detail=str(exc)) from None

    try:
        resp = svc.create_compose_job(
            job_id=job_id,
            face_bytes=raw,
            face_filename=face_image.filename or "face.jpg",
            placement_x=placement_x,
            placement_y=placement_y,
            image_width=image_width,
            image_height=image_height,
            placement_side=side,
            user_prompt=prompt,
            user_id=user_id,
        )
    except Exception as exc:
        raise HTTPException(status_code=500, detail=str(exc)) from None

    return JSONResponse(status_code=202, content=resp.model_dump(mode="json"))


@router.get("/try-on/photo-compose/{compose_id}", response_model=PhotoComposeResponse)
def get_photo_compose(compose_id: str) -> PhotoComposeResponse:
    resp = get_photo_compose_service().get_compose(compose_id)
    if resp is None:
        raise HTTPException(status_code=404, detail="Compose job not found")
    return resp