# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "polyGraphics 3D Backend",
    "description": "API for 2D-to-3D reconstruction. Default path: **Segment Anything first** (smart background removal), then MapAnything / mesh or Gaussian splatting \u2014 reconstructions consume **masked** views only. `skip_sam_segmentation` is only for pre-cut **RGBA** assets with real transparency.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Direct Uvicorn (e.g. http://127.0.0.1:8000)"
    }
  ],
  "paths": {
...
    "/models/{job_id}/try-on": {
      "get": {
        "tags": [
          "try-on"
        ],
        "summary": "Get Model Try On",
        "operationId": "get_model_try_on_models__job_id__try_on_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelTryOnConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "try-on"
        ],
        "summary": "Put Model Try On",
        "operationId": "put_model_try_on_models__job_id__try_on_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelTryOnConfigPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelTryOnConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{job_id}/hanger-point": {
      "get": {
        "tags": [
          "try-on"
        ],
        "summary": "Get Hanger Point",
        "operationId": "get_hanger_point_models__job_id__hanger_point_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelPoint3D"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "try-on"
        ],
        "summary": "Put Hanger Point",
        "operationId": "put_hanger_point_models__job_id__hanger_point_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HangerPointBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelPoint3D"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/try-on-calibration": {
      "get": {
        "tags": [
          "try-on"
        ],
        "summary": "Get User Try On Calibration",
        "operationId": "get_user_try_on_calibration_users_me_try_on_calibration_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTryOnCalibration"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "try-on"
        ],
        "summary": "Put User Try On Calibration",
        "operationId": "put_user_try_on_calibration_users_me_try_on_calibration_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTryOnCalibrationPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTryOnCalibration"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/try-on/photo-compose": {
      "post": {
        "tags": [
          "try-on"
        ],
        "summary": "Post Photo Compose",
        "operationId": "post_photo_compose_try_on_photo_compose_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_photo_compose_try_on_photo_compose_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Compose job accepted; poll GET by compose_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoComposeResponse"
                }
              }
            }
          },
          "200": {
            "description": "Synchronous completion (rare)"
          },
          "503": {
            "description": "AI provider not configured"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/try-on/photo-compose/{compose_id}": {
      "get": {
        "tags": [
          "try-on"
        ],
        "summary": "Get Photo Compose",
        "operationId": "get_photo_compose_try_on_photo_compose__compose_id__get",
        "parameters": [
          {
            "name": "compose_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Compose Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoComposeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_post_photo_compose_try_on_photo_compose_post": {
        "properties": {
          "face_image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Face Image",
            "description": "Side/profile face photo (JPEG or PNG)"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Gallery earring model id"
          },
          "placement_x": {
            "type": "number",
            "title": "Placement X",
            "description": "Normalized X in native image (0=left, 1=right)"
          },
          "placement_y": {
            "type": "number",
            "title": "Placement Y",
            "description": "Normalized Y in native image (0=top, 1=bottom)"
          },
          "image_width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Image Width",
            "description": "Native pixel width of face_image"
          },
          "image_height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Image Height",
            "description": "Native pixel height of face_image"
          },
          "placement_side": {
            "type": "string",
            "title": "Placement Side",
            "description": "left | right | auto",
            "default": "auto"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt",
            "description": "Optional extra instruction (max 2000 chars)"
          }
        },
        "type": "object",
        "required": [
          "face_image",
          "job_id",
          "placement_x",
          "placement_y",
          "image_width",
          "image_height"
        ],
        "title": "Body_post_photo_compose_try_on_photo_compose_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HangerPointBody": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "z": {
            "type": "number",
            "title": "Z"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "z"
        ],
        "title": "HangerPointBody",
        "description": "Backward-compatible hanger-only body (``{ x, y, z }``)."
      },
...
      "JewelryType": {
        "type": "string",
        "enum": [
          "stud",
          "drop"
        ],
        "title": "JewelryType"
      },
...
      "ModelPoint3D": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X",
            "description": "X in normalized try-on space, clamped to [-1, 1]."
          },
          "y": {
            "type": "number",
            "title": "Y",
            "description": "Y in normalized try-on space, clamped to [-1, 1]."
          },
          "z": {
            "type": "number",
            "title": "Z",
            "description": "Z in normalized try-on space, clamped to [-1, 1]."
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "z"
        ],
        "title": "ModelPoint3D",
        "description": "Point in normalized try-on model space (matches frontend ``prepareTryOnModel``).\n\nThe mesh is centered at the origin and the largest axis of its bounding box is 1.0 unit.\nComponents are clamped to [-1, 1] on write."
      },
      "ModelTryOnConfig": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "hanger_point": {
            "$ref": "#/components/schemas/ModelPoint3D"
          },
          "profile_point": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPoint3D"
              },
              {
                "type": "null"
              }
            ]
          },
          "jewelry_rotation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnJewelryRotation"
              },
              {
                "type": "null"
              }
            ]
          },
          "api_vertical_flip": {
            "type": "boolean",
            "title": "Api Vertical Flip",
            "default": false
          },
          "ear_offsets": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnEarOffsets"
              },
              {
                "type": "null"
              }
            ]
          },
          "side_lateral_nudge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnSideLateralNudge"
              },
              {
                "type": "null"
              }
            ]
          },
          "jewelry_type": {
            "$ref": "#/components/schemas/JewelryType",
            "default": "drop"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "UTC timestamp ISO-8601 (e.g. 2026-06-02T12:00:00Z)."
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "hanger_point",
          "updated_at"
        ],
        "title": "ModelTryOnConfig"
      },
      "ModelTryOnConfigPut": {
        "properties": {
          "hanger_point": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPoint3D"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile_point": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPoint3D"
              },
              {
                "type": "null"
              }
            ]
          },
          "jewelry_rotation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnJewelryRotation"
              },
              {
                "type": "null"
              }
            ]
          },
          "api_vertical_flip": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Vertical Flip"
          },
          "ear_offsets": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnEarOffsets"
              },
              {
                "type": "null"
              }
            ]
          },
          "side_lateral_nudge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnSideLateralNudge"
              },
              {
                "type": "null"
              }
            ]
          },
          "jewelry_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JewelryType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ModelTryOnConfigPut",
        "description": "Upsert body. Omitted fields are merged with the existing row on PUT."
      },
      "PhotoComposeResponse": {
        "properties": {
          "compose_id": {
            "type": "string",
            "title": "Compose Id"
          },
          "status": {
            "$ref": "#/components/schemas/PhotoComposeStatus"
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          },
          "result_image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Image Base64"
          },
          "revised_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revised Prompt"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "compose_id",
          "status"
        ],
        "title": "PhotoComposeResponse"
      },
...
      "TryOnEarOffsets": {
        "properties": {
          "vertical": {
            "type": "number",
            "title": "Vertical",
            "description": "Fine-tune vertical placement on face.",
            "default": 0.0
          },
          "depth": {
            "type": "number",
            "title": "Depth",
            "description": "Fine-tune depth placement on face.",
            "default": 0.0
          },
          "lateral": {
            "type": "number",
            "title": "Lateral",
            "description": "Fine-tune lateral placement on face.",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "TryOnEarOffsets"
      },
      "TryOnJewelryRotation": {
        "properties": {
          "spin_deg": {
            "type": "number",
            "title": "Spin Deg",
            "description": "Yaw around hanger, degrees [-180, 180].",
            "default": 0.0
          },
          "pitch_deg": {
            "type": "number",
            "title": "Pitch Deg",
            "description": "Pitch, degrees [-75, 75].",
            "default": 0.0
          },
          "roll_deg": {
            "type": "number",
            "title": "Roll Deg",
            "description": "Roll, degrees [-180, 180].",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "TryOnJewelryRotation"
      },
...
      "UserTryOnCalibration": {
        "properties": {
          "left_offset": {
            "$ref": "#/components/schemas/TryOnEarOffsets"
          },
          "right_offset": {
            "$ref": "#/components/schemas/TryOnEarOffsets"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "left_offset",
          "right_offset",
          "updated_at"
        ],
        "title": "UserTryOnCalibration"
      },
      "UserTryOnCalibrationPut": {
        "properties": {
          "left_offset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnEarOffsets"
              },
              {
                "type": "null"
              }
            ]
          },
          "right_offset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TryOnEarOffsets"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "UserTryOnCalibrationPut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Google sign-in access token from POST /auth/google or GET /auth/google/callback."
      }
    }
  }
}